+++ data-raw/iris-fst.R
@@ -1,0 +1,10 @@
+#################################################################%#
+#### Code to create the rds file `iris.fst sous `inst/extdata`####
+###############################################################%#
+
+library(fst)
+
+data(iris)
+
+fst::write.fst(x = iris,
+               path = "inst/extdata/iris.fst")